Add files via upload
This is 'Automated Data Cleaning Project' with MySQL.
I created a stored procedure that automatically cleans the data for us (remove duplicates & standardizes data). Then, I created an event that runs the stored procedure at a scheduled time.
I also added a debugging section that checks if the procedure worked.